pyproject.toml
[project]
name = "poetry-demo"
version = "0.1.0"
description = "Document Analysis API with FastAPI, LlamaIndex, and Redis"
authors = [
    {name = "hyunbae.jeon", email = "mazingazz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn[standard]>=0.35.0,<0.36.0",
    "oracledb>=3.2.0,<4.0.0",

    "sqlalchemy>=2.0.42,<3.0.0",
    "sqlalchemy-utils==0.41.1",
    "pydantic>=2.11.7,<3.0.0",
    "pandas>=2.3.1,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "orjson>=3.11.1,<4.0.0",
    "langchain[openai]>=0.3.27,<0.4.0",
    "langchain-openai>=0.3.30,<0.4.0",
    "langchain-community>=0.3.27,<0.4.0",
    "langchain-core>=0.3.74,<0.4.0",
    "langsmith>=0.3.45,<0.4.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "openai>=1.101.0,<2.0.0",
    "bs4>=0.0.2,<0.0.3",
    "faiss-cpu>=1.12.0,<2.0.0",
    "sentence-transformers>=5.1.0,<6.0.0",
    "mcp>=1.3.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "loguru>=0.7.0,<0.8.0",
    "llama-index>=0.12.0,<0.13.0",
    "pymupdf>=1.24.0,<2.0.0",
    "redis>=5.0.0,<6.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.14.0",
]

[project.scripts]
start = "app.main:start"
mcp-server = "app.mcp_server:main"
mcp-example = "app.examples.mcp_client_example:main"
async-patterns = "app.examples.async_patterns:main"
lcel-patterns = "app.examples.lcel_patterns:main"
llamaindex-patterns = "app.examples.llamaindex_patterns:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pydantic-settings>=2.10.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.14.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
